+++ salesmanago/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.11.0 released 26.03.2026
+- Added support array-based opt-in checkbox values in forms
+- Fixed issue with duplicated events when order status changes
+- Added woocommerce_thankyou hook handling for external events
+- 
 # 3.10.0 released 25.02.2026
 - Added a question to the FAQ regarding security
 - Tiered Pricing plugin support added